fix url in readme (#24)
The URL now has the correct repo and fingerprint.

The qr code in the repo itself, shown when you open it in fdroid, is generated automatically and won't display the same way.

The qr code in the readme doesn't include the fingerprint. I can regenerate that if needed, but it doesn't seem necessary.